♻️ Refactor search component
